Database: refactor to fix neon connection-dropping problem
So, [neon](https://neon.com/) (a managed postgres paas) has a model
where after 5 minutes of inactivity, it will terminate your connections
and put the database into an idle state. (I want to try it for kadreg
because that model makes it very cheap as a managed postgres option
for smaller projects).

This seems to interact poorly with pgo's pooling model. (pgo being the
erlang library used by pog). Pgo starts a pool of connections at the
full pool size on start and expects to keep them all alive permanently.

Neon has a post about [using neon with long-running
applications](https://neon.com/blog/using-neons-auto-suspend-with-long-running-applications),
but it assumes a model of connection pooling where the pool does not
expect to keep all connections alive. A pool that either closes idle
connections itself after some configurable timeout, or at least expects
that if a connection disconnects, it just emits a warning and shrugs,
and will start a new connection when smoeone next makes a query.

And pgo doesn't work that way. For a start it has a default 1s
heartbeat on all connections, which keeps the neon db alive, but
that's easily solvable by setting idle_interval to something very
large. But the main issue is just that it tries to keep all its
pool connections alive, reconnecting if they close. (In fact that
doesn't seem to work, and when neon disconnects all the
connections it just starts giving QueryTimeout errors to all
subsequent queries, but shrug, that's no better. also can't see
any way to get observability of this).

AFAICS my options are:
1. find a different erlang or elixir postgres client that has
behaviour more compatible with neon's model, and write some gleam
bindings for it. (i can't find any non-pgo-based gleam libs).
2. use pog but with an abstraction layer, where I track the time of
the last query and if it's been over two minutes I assume the
database might have gone idle, kill the pool process and start a new
one.
3. forget neon and migrate to a more traditional managed postgres.

This commit implements option 2.

All queries are funneled through a genserver, which tracks the last
query time and decides whether to recycle the conn appropriately.